Moves updating Graph DB to the GraphSelectGuidancePlugin. Closes #481
